Move impl Evaluate for Instance to v1_ext/instance.rs
