+++ rust/ommx/src/evaluate.rs
@@ -1,9 +1,6 @@
-use crate::v1::{
-    Function, Instance, Optimality, Relaxation, SampleSet, SampledDecisionVariable, Samples,
-    Solution, State,
-};
-use anyhow::{bail, Result};
-use std::collections::{hash_map::Entry as HashMapEntry, BTreeSet, HashMap};
+use crate::v1::{Samples, State};
+use anyhow::Result;
+use std::collections::BTreeSet;
 
 /// Evaluate with a [State]
 pub trait Evaluate {
@@ -23,207 +20,16 @@ pub trait Evaluate {
     fn required_ids(&self) -> BTreeSet<u64>;
 }
 
-impl Evaluate for Instance {
-    type Output = Solution;
-    type SampledOutput = SampleSet;
-
-    fn evaluate(&self, state: &State) -> Result<Self::Output> {
-        self.check_bound(state, 1e-7)?;
-        let mut evaluated_constraints = Vec::new();
-        let mut feasible_relaxed = true;
-        for c in &self.constraints {
-            let c = c.evaluate(state)?;
-            // Only check non-removed constraints for feasibility
-            if feasible_relaxed {
-                feasible_relaxed = c.is_feasible(1e-6)?;
-            }
-            evaluated_constraints.push(c);
-        }
-        let mut feasible = feasible_relaxed;
-        for c in &self.removed_constraints {
-            let c = c.evaluate(state)?;
-            if feasible {
-                feasible = c.is_feasible(1e-6)?;
-            }
-            evaluated_constraints.push(c);
-        }
-
-        let objective = self.objective().evaluate(state)?;
-
-        let mut state = state.clone();
-        for v in &self.decision_variables {
-            if let Some(value) = v.substituted_value {
-                state.entries.insert(v.id, value);
-            }
-        }
-        eval_dependencies(&self.decision_variable_dependency, &mut state)?;
-        for v in &self.decision_variables {
-            if let HashMapEntry::Vacant(e) = state.entries.entry(v.id) {
-                let bound: crate::Bound = v.try_into()?;
-                e.insert(bound.nearest_to_zero());
-            }
-        }
-        Ok(Solution {
-            decision_variables: self.decision_variables.clone(),
-            state: Some(state),
-            evaluated_constraints,
-            feasible_relaxed: Some(feasible_relaxed),
-            feasible,
-            objective,
-            optimality: Optimality::Unspecified.into(),
-            relaxation: Relaxation::Unspecified.into(),
-            ..Default::default()
-        })
-    }
-
-    fn partial_evaluate(&mut self, state: &State) -> Result<()> {
-        for v in &mut self.decision_variables {
-            if let Some(value) = state.entries.get(&v.id) {
-                v.substituted_value = Some(*value);
-            }
-        }
-        if let Some(f) = self.objective.as_mut() {
-            f.partial_evaluate(state)?
-        }
-        for constraints in &mut self.constraints {
-            constraints.partial_evaluate(state)?;
-        }
-        for constraints in &mut self.removed_constraints {
-            constraints.partial_evaluate(state)?;
-        }
-        for d in self.decision_variable_dependency.values_mut() {
-            d.partial_evaluate(state)?;
-        }
-        Ok(())
-    }
-
-    fn evaluate_samples(&self, samples: &Samples) -> Result<Self::SampledOutput> {
-        let mut feasible_relaxed: HashMap<u64, bool> =
-            samples.ids().map(|id| (*id, true)).collect();
-
-        // Constraints
-        let mut constraints = Vec::new();
-        for c in &self.constraints {
-            let evaluated = c.evaluate_samples(samples)?;
-            for (sample_id, feasible_) in evaluated.is_feasible(1e-6)? {
-                if !feasible_ {
-                    feasible_relaxed.insert(sample_id, false);
-                }
-            }
-            constraints.push(evaluated);
-        }
-        let mut feasible = feasible_relaxed.clone();
-        for c in &self.removed_constraints {
-            let v = c.evaluate_samples(samples)?;
-            for (sample_id, feasible_) in v.is_feasible(1e-6)? {
-                if !feasible_ {
-                    feasible.insert(sample_id, false);
-                }
-            }
-            constraints.push(v);
-        }
-
-        // Objective
-        let objectives = self.objective().evaluate_samples(samples)?;
-
-        // Reconstruct decision variable values
-        let mut samples = samples.clone();
-        for state in samples.states_mut() {
-            eval_dependencies(&self.decision_variable_dependency, state?)?;
-        }
-        let mut transposed = samples.transpose();
-        let decision_variables: Vec<SampledDecisionVariable> = self
-            .decision_variables
-            .iter()
-            .map(|d| -> Result<_> {
-                Ok(SampledDecisionVariable {
-                    decision_variable: Some(d.clone()),
-                    samples: transposed.remove(&d.id),
-                })
-            })
-            .collect::<Result<_>>()?;
-
-        Ok(SampleSet {
-            decision_variables,
-            objectives: Some(objectives),
-            constraints,
-            feasible_relaxed,
-            feasible,
-            sense: self.sense,
-            ..Default::default()
-        })
-    }
-
-    fn required_ids(&self) -> BTreeSet<u64> {
-        self.used_decision_variable_ids()
-    }
-}
-
-// FIXME: This would be better by using a topological sort
-fn eval_dependencies(dependencies: &HashMap<u64, Function>, state: &mut State) -> Result<()> {
-    let mut bucket: Vec<_> = dependencies.iter().collect();
-    let mut last_size = bucket.len();
-    let mut not_evaluated = Vec::new();
-    loop {
-        while let Some((id, f)) = bucket.pop() {
-            match f.evaluate(state) {
-                Ok(value) => {
-                    state.entries.insert(*id, value);
-                }
-                Err(_) => {
-                    not_evaluated.push((id, f));
-                }
-            }
-        }
-        if not_evaluated.is_empty() {
-            return Ok(());
-        }
-        if last_size == not_evaluated.len() {
-            bail!("Cannot evaluate any dependent variables.");
-        }
-        last_size = not_evaluated.len();
-        bucket.append(&mut not_evaluated);
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::{
         random::*,
-        v1::{Linear, Polynomial, Quadratic},
+        v1::{Function, Instance, Linear, Polynomial, Quadratic},
     };
     use approx::*;
     use maplit::*;
     use proptest::prelude::*;
-
-    #[test]
-    fn test_eval_dependencies() {
-        let mut state = State::from_iter(vec![(1, 1.0), (2, 2.0), (3, 3.0)]);
-        let dependencies = hashmap! {
-            4 => Function::from(Linear::new([(1, 1.0), (2, 2.0)].into_iter(), 0.0)),
-            5 => Function::from(Linear::new([(4, 1.0), (3, 3.0)].into_iter(), 0.0)),
-        };
-        eval_dependencies(&dependencies, &mut state).unwrap();
-        assert_eq!(state.entries[&4], 1.0 + 2.0 * 2.0);
-        assert_eq!(state.entries[&5], 1.0 + 2.0 * 2.0 + 3.0 * 3.0);
-
-        // circular dependency
-        let mut state = State::from_iter(vec![(1, 1.0), (2, 2.0), (3, 3.0)]);
-        let dependencies = hashmap! {
-            4 => Function::from(Linear::new([(1, 1.0), (5, 2.0)].into_iter(), 0.0)),
-            5 => Function::from(Linear::new([(4, 1.0), (3, 3.0)].into_iter(), 0.0)),
-        };
-        assert!(eval_dependencies(&dependencies, &mut state).is_err());
-
-        // non-existing dependency
-        let mut state = State::from_iter(vec![(1, 1.0), (2, 2.0), (3, 3.0)]);
-        let dependencies = hashmap! {
-            4 => Function::from(Linear::new([(1, 1.0), (6, 2.0)].into_iter(), 0.0)),
-            5 => Function::from(Linear::new([(4, 1.0), (3, 3.0)].into_iter(), 0.0)),
-        };
-        assert!(eval_dependencies(&dependencies, &mut state).is_err());
-    }
 
     #[test]
     fn linear_partial_evaluate() {
